Move the form related into its own file
